added feature to report metrics of aliased indexes
if you specify 'alias' along with the 'indices' in the type of things to index, only aliased indexes will be reported.
this is a feature for people using aliases to atomicaly swap indices.
use case is :
- many instances of an application which could update an index atomicaly
- one instance will create a new index prefixed with a serial number (like the current miliseconds since 1970) ex : "index-application-12345678"
- the instance then atomicaly create/change an alias "index-application" to point to index "index-application-12345678"

this is the same issue as with the logstash index creating one index per day. 
there is already a solution to "remove" the date part of the index.
my solution is on ES side : for logstash, you could create an alias that would include every daily index, then have your dashboard querying this alias. Also, you only want an aggregation of the stats, not the stat of each index individualy.

This patch is simple :
if the "alias" keyword is in the config, then, when parsing the index metrics, the index path is looked up agains the alias list. If the name is in an alias, values are sum and path is changed to the alias path. If not, indice path is not parsed.